Refactor fee naming pattern to match the shipping one
